web/scantact/requirements.txt
Babel==2.2.0
Django==1.9.2
django-filter==0.12.0
django-phonenumber-field==1.0.0
djangorestframework==3.3.2
Markdown==2.6.5
phonenumbers==7.2.5
Pillow==3.1.1
pytz==2015.7
